# server/openapi.yaml
openapi: 3.0.3
info:
  title: NetTechPro API
  description: |
    RESTful API cho website bán thiết bị mạng NetTechPro.
    
    ## Tính năng
    - Quản lý sản phẩm (Products)
    - Quản lý danh mục (Categories)
    - Quản lý thương hiệu (Brands)
    - Quản lý đơn hàng (Orders)
    - Giỏ hàng (Cart)
    - Xác thực (Authentication)
    
    ## Mock Data
    - 8 sản phẩm thiết bị mạng
    - 8 danh mục
    - 10 thương hiệu
    - 3 đơn hàng mẫu
  version: 1.0.0
  contact:
    name: NetTechPro API Support
    email: support@nettechpro.me
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.nettechpro.me
    description: Production server
...
paths:
...
  /api/orders:
    get:
      summary: Lấy tất cả đơn hàng (Admin)
      tags: [Orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: Danh sách đơn hàng
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
    post:
      summary: Tạo đơn hàng mới
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Đơn hàng đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/my:
    get:
      summary: Lấy đơn hàng của tôi
      tags: [Orders]
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Danh sách đơn hàng
  /api/orders/{id}:
    get:
      summary: Lấy chi tiết đơn hàng
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết đơn hàng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/{id}/status:
    put:
      summary: Cập nhật trạng thái đơn hàng
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: Trạng thái đã được cập nhật
  /api/orders/{id}/cancel:
    post:
      summary: Hủy đơn hàng
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Đơn hàng đã được hủy
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
          example: ORD-20251015-001
        userId:
          type: string
        userName:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              productName:
                type: string
              productNameVi:
                type: string
              quantity:
                type: integer
              price:
                type: number
        subtotal:
          type: number
        shippingFee:
          type: number
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        statusVi:
          type: string
        paymentMethod:
          type: string
          enum: [cod, bank_transfer, momo, vnpay, credit_card]
        paymentMethodVi:
          type: string
        shippingAddress:
          type: object
          properties:
            fullName:
              type: string
            phone:
              type: string
            address:
              type: string
            city:
              type: string
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [items, shippingAddress, paymentMethod]
      properties:
        userId:
          type: string
        userName:
          type: string
        items:
          type: array
          items:
            type: object
        shippingAddress:
          type: object
        paymentMethod:
          type: string
...